+++ examples/field_only/source_absorb.input
