+++ examples/full/bernstein.input
