+++ examples/full/bernstein_main.cpp
