+++ examples/single_particle/long_stability.input
