+++ examples/single_particle/osc.input
