+++ examples/single_particle/osc_elec.input
