+++ include/particle_defs.hpp
@@ -25,9 +25,6 @@
 #define C_NUM_SOA_INTS 0 
 
 
-#ifndef WRANGE
-#define WRANGE 2
-#endif
 
 
 // Pos variable indices
