+++ include/w_defs.hpp
@@ -11,6 +11,7 @@ amrex::Real W1(amrex::Real x);
 amrex::Real Wp(amrex::Real x);
 amrex::Real I_W1(amrex::Real a,amrex::Real b);
 amrex::Real I_Wp(amrex::Real a,amrex::Real b);
+extern const int interpolation_range;
 
 
 #endif
