+++ src/interpolation/Makefile
@@ -1,0 +1,11 @@
+SRCS    := $(wildcard *.cpp)
+OBJS    := $(patsubst %.cpp,%.o,$(SRCS))
+
+all: $(OBJS)
+
+%.o : %.cpp %.hpp
+	$(CXX) $(CXXFLAGS) $< -o $@ -c
+
+.PHONY: clean
+clean:
+	rm -f *.o
