+++ tests/energy_conservation/energy.input
