+++ tests/energy_conservation/energy_other.input
