+++ tests/single_particle/break.input
