+++ tests/single_particle/cyclotron_borders.input
